Some additions
I got rid of the side bar as it's not needed. I also deleted a file that was un needed. I have made a rough draft of the account look and am currently thinking of a way to implament a "follow" system after I figure out the way users can make and share drinks as well as games.